Mobile_net_v2_complete_project_front_back1/BrainTumorApp/backend1/requirements.txt
Django>=4.2
djangorestframework
tensorflow
pillow
django-cors-headers
